+++ vite.config.js
@@ -2,7 +2,7 @@ import { defineConfig } from 'vite';
 import { resolve } from 'node:path';
 
 export default defineConfig({
-  root: resolve(process.cwd(), 'app'),
+  root: resolve(process.cwd(), 'product/app'),
   server: {
     fs: {
       allow: [resolve(process.cwd())]